jointest for flipscores objects

jointest for flipscores objects

join_flipscores(
  mods,
  tested_coeffs = NULL,
  n_flips = 5000,
  score_type = "standardized",
  statistics = "t",
  seed = NULL,
  ...
)

Arguments

mods

list of glms (or list of any other object that can be evaluated by flipscores)

tested_coeffs

is a list of the same length of mods, each element of the list being a vector of of names of tested coefficients. Alternatively, it can be a vector of names of tested coefficients, in this case, the tested coefficients are attributed to all models (when present). As a last option, it can be NULL, if so, all coefficients are tested.

n_flips

= 5000

score_type

any valid type for flipscores, "standardized" is the default. see flipscores for more datails

statistics

"t" is the only method implemented (yet). Any other value will not modify the Score (a different statistic will only affect the multivariate inference, not the univariate one).

seed

NULL by default. Use a number if you wanto to ensure replicability of the results

...

any other further parameter.

Examples

n=20
set.seed(123)
D=data.frame(X=rnorm(n),Z1=rnorm(n),Z2=rnorm(n))
D$Y=D$Z1+D$X+rnorm(n)
mod1=glm(Y~X+Z1+Z2,data=D)
mod2=glm(Y~X+poly(Z1,2)+Z2,data=D)
mod3=glm(Y~X+poly(Z1,2)+poly(Z2,2),data=D)
mod4=glm(Y~X+Z1+poly(Z2,2),data=D)
mods=list(mod1=mod1,mod2=mod2,mod3=mod3,mod4=mod4)
for(i in 1:length(mods))
mods[[i]]$call$data=eval(D)
library(jointest)
res=join_flipscores(mods,n_flips = 5000, score_type = "standardized" ,
      seed = 1, tested_coeffs = "X")
summary(res)
#>      Coeff  Estimate    Score Std. Error  z value Part. Cor   Pr(>|t|)
#> mod1     X 1.1254046 19.78375   6.309305 3.135646 0.7605058 0.01439712
#> mod2     X 0.9552644 15.38036   5.053879 3.043278 0.7608195 0.01539692
#> mod3     X 1.0121921 14.31866   4.836669 2.960437 0.7643816 0.01579684
#> mod4     X 1.1696906 17.62979   5.875380 3.000621 0.7501552 0.01259748
summary(combine(res))
#>   Coeff Stat nMods        S          p
#> X     X maxT     4 19.78375 0.01559688
res=p.adjust.fwer(res)
summary(res)
#>      Coeff  Estimate    Score Std. Error  z value Part. Cor   Pr(>|t|)
#> mod1     X 1.1254046 19.78375   6.309305 3.135646 0.7605058 0.01439712
#> mod2     X 0.9552644 15.38036   5.053879 3.043278 0.7608195 0.01539692
#> mod3     X 1.0121921 14.31866   4.836669 2.960437 0.7643816 0.01579684
#> mod4     X 1.1696906 17.62979   5.875380 3.000621 0.7501552 0.01259748
#>           p.adj
#> mod1 0.01559688
#> mod2 0.01939612
#> mod3 0.01939612
#> mod4 0.01559688