Correct p-values of a jointest object for multiple testing (FamilyWise Error Rate)

Correct p-values of a jointest object for multiple testing (FamilyWise Error Rate)

p.adjust.fwer(mods, method = "maxT", tail = 0, ...)

Arguments

mods

object of class jointest (it can be a list of glm or flipscores converted in a jointest object using as.jointest)

method

any method implemented in flip::flip.adjust or a custom function. In the last case it must be a function that uses a matrix as input and returns a vector of adjusted p-values equal to the numbero fo columns of the inputed matrix.

tail

= 0

...

further parameters